Windows CLI Improvement
The Ansi parser and their associated actions have been decoupled. Now
parsing results in call backs to an interface which performs the
appropriate actions depending on the environment.  Right now there is a
test implementation and an initial Windows implementation.

This improvement provides a functional Vi experience and the vttest no
longer panics.

Signed-off-by: Gabriel Hartmann <gabhart@microsoft.com>